Fix layout and add listeners to populate display
The auxiliary button layout was broken after changing the flow direction
of the buttons div to column. It has been fixed by fixing the space in
between each button using a gap of 10px.

Every button now has an event listener attached to it. If a user clicks
a button and it is a number, it populates the display screen with their
input.