Add functionality to delete, ., and clear buttons
The delete button slices the display text to remove one digit each time
the button is pressed.

The all clear button clears the display text entirely to allow the user
to start anew.

The decimal button allows the user to place one single decimal in a
number. If there is already a decimal in the display text, the button
won't work.