Add logic for evaluating individual expressions
The calculator can now evaluate individual expressions, like 3 + 2 = 5.
It does this by using an array that holds three values, in this order:
 * The first operand
 * The second operand
 * The operator (*, /, +, -)

When the user inputs a number and hits an operator, those two values
are stored in currentOperation[0] (the first number) and
currentOperation[2] (the operator).

When equals is pressed, if there is already one operand present in the
array, it populates the array with the rest of the values (the second
operand), and evaluates the expression using the operate() function.