Allow expression evaluation after user hits = sign
Previously, the calculator was unable to continue evaluating an
expression from the result achieved by pressing the equals (=) sign.

For instance, if a user entered 6 * 3 =, the calculator would return 18,
but the user would not be able to continue working with that number. For
example, trying to add two to it: 6 * 3 = 18 + 2 would not work.

Additionally, the calculator used to break if the user tried to select
multiple operators. Now, the only operator that is saved is the last one
that the user pressed. For example, if they enter 5 * - / + 3, the
calculator will register this as 5 + 3.