Fix display issue caused after operand evaluation
The display would only show single digits at a time after the operand
was registered.

For example, if the user attempted the following:
    3 * 63312

The display would only show the number 2 (the last single digit in the
number).

This was caused by a logical error within the button click event
listener. Each time a button was pressed, the display would clear if
there was a first operand already stored in the currentExpression array.