Improve boolean logic to account for more cases
The debounce boolean made to fix the issue of only single digits being
displayed after an operator was registered was renamed to
"resultOnScreen" and changed to be true if the number on the display
is the result of a operation. This allows us to expand the use of this
boolean to more cases.

For instance, a user was able to delete digits from the reuslt of an
operation. Now, the user is only able to delete digits from numbers they
have entered themselves.

Additionally, it is more intuitive for the screen to clear when the
user presses a button only if the numbers currently on screen are the
result of an operation.