Fix evaluating from 0 failure
If the user tried to evaluate an expression starting from 0, i.e. 0 - 2,
the calculator would fail. This happened because the logic used to
evaluate the operation was equating 0 to a falsy value, causing it
to skip over necessary steps.

This was fixed by using !isNan(currentExpression[0]) to ensure the
existence of a number in the array is being properly checked.