Fix decimal point failure in chained operations
If the user were to evaluate an expression with chained decimal numbers
without a zero preceding them, i.e.:
    6 * 3 * .3 - .2
The calculator would fail. This happened because of an error in logic
originally intended to make sure the user could not type in several
decimal points within one number, for example, 6.23.1.

Old logic checked whether there was already a decimal point present in
the number currently being shown on the display.

However, this was flawed because if the number currently being shown on
the display was the result of a previous operation, and it included a
decimal point, the program would still register that number as user
input and disable the decimal digit for safety.

This was fixed by changing the logic to do the following:
 * First, check for a decimal point.
 * If one is present, determine whether it is from a previous operation.
 * If it is the result of an operation, clear the screen and allow
   decimal input.
 * If it is the result of user input, do not register the decimal input.